Fix delete methods. Refactored duplicated code segments. Cleanup

+++ src/main/java/com/dyalex/personalfinancemephi/service/TransactionService.java
@@ -5,6 +5,8 @@ import com.dyalex.personalfinancemephi.model.Transaction;
 import com.dyalex.personalfinancemephi.model.TransactionType;
 import com.dyalex.personalfinancemephi.model.Wallet;
 import com.dyalex.personalfinancemephi.repository.TransactionRepository;
+import com.dyalex.personalfinancemephi.repository.WalletRepository;
+import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -20,6 +22,7 @@ import java.util.stream.Collectors;
 public class TransactionService {
     private final TransactionRepository transactionRepository;
     private final WalletService walletService;
+    private final WalletRepository walletRepository;
 
     public List<Transaction> getAllTransactions(Long userId) {
         List<Wallet> wallets = walletService.getAllWallets(userId);
@@ -35,8 +38,27 @@ public class TransactionService {
         return transactionRepository.findById(id).orElse(null);
     }
 
+    @Transactional
     public void deleteTransactionById(UUID id) {
-        transactionRepository.deleteById(id);
+        Transaction transaction = transactionRepository.findById(id).orElse(null);
+        if (transaction != null) {
+            TransactionType type = transaction.getCategory().getType();
+            Wallet wallet = transaction.getWallet();
+            switch (type) {
+                case EXPENSE:
+                    wallet.setBalance(wallet.getBalance().add(transaction.getAmount()));
+                    walletRepository.save(wallet);
+                    transactionRepository.deleteById(id);
+                    return;
+                case INCOME:
+                    wallet.setBalance(wallet.getBalance().subtract(transaction.getAmount()));
+                    walletRepository.save(wallet);
+                    transactionRepository.deleteById(id);
+                    return;
+                default:
+                    transactionRepository.deleteById(id);
+            }
+        }
     }
 
     public void saveExpense(Transaction expense) {
@@ -60,15 +82,9 @@ public class TransactionService {
     }
 
     public void updateExpense(UUID id, Transaction expense) {
-        Transaction transaction = transactionRepository.findById(id).orElse(null);
-        if (transaction != null) {
-            Transaction newTransaction = Transaction.builder()
-                    .id(transaction.getId())
-                    .amount(expense.getAmount() != null ? expense.getAmount() : transaction.getAmount())
-                    .date(expense.getDate() != null ? expense.getDate() : LocalDate.now())
-                    .category(expense.getCategory() != null ? expense.getCategory() : transaction.getCategory())
-                    .wallet(expense.getWallet() != null ? expense.getWallet() : transaction.getWallet())
-                    .build();
+        Transaction oldTransaction = transactionRepository.findById(id).orElse(null);
+        if (oldTransaction != null) {
+            Transaction newTransaction = buildNewTransaction(oldTransaction, expense);
 
             Wallet wallet = walletService.getWalletById(expense.getWallet().getId());
             if (wallet != null) {
@@ -101,15 +117,9 @@ public class TransactionService {
     }
 
     public void updateIncome(UUID id, Transaction income) {
-        Transaction transaction = transactionRepository.findById(id).orElse(null);
-        if (transaction != null) {
-            Transaction newTransaction = Transaction.builder()
-                    .id(transaction.getId())
-                    .amount(income.getAmount() != null ? income.getAmount() : transaction.getAmount())
-                    .date(income.getDate() != null ? income.getDate() : LocalDate.now())
-                    .category(income.getCategory() != null ? income.getCategory() : transaction.getCategory())
-                    .wallet(income.getWallet() != null ? income.getWallet() : transaction.getWallet())
-                    .build();
+        Transaction oldTransaction = transactionRepository.findById(id).orElse(null);
+        if (oldTransaction != null) {
+            Transaction newTransaction = buildNewTransaction(oldTransaction, income);
             Wallet wallet = walletService.getWalletById(income.getWallet().getId());
             if (wallet != null) {
                 wallet.setBalance(wallet.getBalance().add(income.getAmount()));
@@ -117,5 +127,15 @@ public class TransactionService {
                 walletService.saveWallet(wallet);
             }
         }
+    }
+
+    private Transaction buildNewTransaction(Transaction oldTransaction, Transaction transaction) {
+        return Transaction.builder()
+                .id(oldTransaction.getId())
+                .amount(transaction.getAmount() != null ? transaction.getAmount() : oldTransaction.getAmount())
+                .date(transaction.getDate() != null ? transaction.getDate() : LocalDate.now())
+                .category(transaction.getCategory() != null ? transaction.getCategory() : oldTransaction.getCategory())
+                .wallet(transaction.getWallet() != null ? transaction.getWallet() : oldTransaction.getWallet())
+                .build();
     }
 }
